refactor: decouple the libretro core from the game
