+++ cmake/ncine_extra_sources.cmake
@@ -566,6 +566,13 @@ endif()
 #	list(APPEND SOURCES ${NCINE_SOURCE_DIR}/nCine/Graphics/RenderDocCapture.cpp)
 #endif()
 
+if(NOT CURL_FOUND AND NOT WIN32)
+	# `WebRequest` needs an HTTP backend: WinHTTP on Windows, libcurl everywhere else. Without one
+	# it cannot be compiled at all, so it's excluded together with everything that depends on it
+	# (the update check and the online server list, see `WITH_ONLINE_MULTIPLAYER`).
+	list(REMOVE_ITEM SOURCES "${NCINE_SOURCE_DIR}/Shared/IO/WebRequest.cpp")
+endif()
+
 if(CURL_FOUND)
 	target_compile_definitions(${NCINE_APP} PRIVATE "WITH_CURL")
 	target_link_libraries(${NCINE_APP} PRIVATE CURL::libcurl)
@@ -717,9 +724,15 @@ elseif(WINDOWS_PHONE OR WINDOWS_STORE)
 else()
 	if(NCINE_BUILD_LIBRETRO)
 		# The libretro entry point replaces MainApplication (the frontend drives the game loop)
-		target_compile_definitions(${NCINE_APP} PRIVATE "WITH_LIBRETRO" "NCINE_PACKAGED_CONTENT_PATH")
+		target_compile_definitions(${NCINE_APP} PRIVATE "WITH_LIBRETRO" "NCINE_PACKAGED_CONTENT_PATH" "NCINE_APP=\"${NCINE_APP}\"")
 		list(APPEND HEADERS ${NCINE_SOURCE_DIR}/libretro/libretro.h)
+		list(APPEND HEADERS ${NCINE_SOURCE_DIR}/nCine/Backends/LibretroApplication.h)
+		list(APPEND HEADERS ${NCINE_SOURCE_DIR}/nCine/Backends/LibretroGfxDevice.h)
+		list(APPEND HEADERS ${NCINE_SOURCE_DIR}/nCine/Backends/LibretroInputManager.h)
 		list(APPEND SOURCES ${NCINE_SOURCE_DIR}/libretro.cpp)
+		list(APPEND SOURCES ${NCINE_SOURCE_DIR}/nCine/Backends/LibretroApplication.cpp)
+		list(APPEND SOURCES ${NCINE_SOURCE_DIR}/nCine/Backends/LibretroGfxDevice.cpp)
+		list(APPEND SOURCES ${NCINE_SOURCE_DIR}/nCine/Backends/LibretroInputManager.cpp)
 	else()
 		list(APPEND HEADERS ${NCINE_SOURCE_DIR}/nCine/MainApplication.h)
 		list(APPEND SOURCES ${NCINE_SOURCE_DIR}/nCine/MainApplication.cpp)
